moved website content into frontend directory
When managing and serving content with git, best practice is to
reserve the root of the repository for READMEs, .git, and other
files that are best not made visible on a live website.

It's particularly important to make sure .git isn't served, for
reasons that should be obvious. This *can* be accomplished using
access controls on the web server, but this way is cleaner and
more maintainable.